+++ demo/src/sandboxes/leva-minimal/src/App.jsx
@@ -6,7 +6,12 @@ export default function App() {
   const data = useControls({
     number: 10,
     minmax: { value: 12.5, min: 5.5, max: 30.5, optional: true },
-    color: { value: '#f00', hint: 'Hey, we support icons and hinting values!', label: <Half2Icon /> },
+    printSize: { value: 100, min: 80, max: 140, step: 10 },
+    color: {
+      value: '#f00',
+      hint: 'Hey, we support icons and hinting values and long text will wrap!',
+      label: <Half2Icon />,
+    },
   })
 
   return (