+++ packages/leva/src/components/ButtonGroup/StyledButtonGroup.tsx
@@ -1,7 +1,7 @@
 import { styled } from '../../styles'
 
 export const StyledButtonGroup = styled('div', {
-  $leva__flex: '',
+  $flex: '',
   justifyContent: 'flex-end',
-  gap: '$leva__colGap',
+  gap: '$colGap',
 })