+++ packages/leva/src/components/Leva/Filter.tsx
@@ -7,9 +7,9 @@ import { StyledFilterInput, StyledTitleWithFilter, TitleContainer, Icon, FilterW
 import { useStoreContext } from '../../context'
 import { DataInput } from '../../types'
 
-type FilterProps = { setFilter: (value: string) => void }
+type FilterProps = { setFilter: (value: string) => void; toggle: (flag?: boolean) => void }
 
-const FilterInput = React.forwardRef<HTMLInputElement, FilterProps>(({ setFilter }, ref) => {
+const FilterInput = React.forwardRef<HTMLInputElement, FilterProps>(({ setFilter, toggle }, ref) => {
   const [value, set] = useState('')
   const debouncedOnChange = useMemo<FilterProps['setFilter']>(() => debounce(setFilter, 250), [setFilter])
   const clear = () => {
@@ -19,6 +19,7 @@ const FilterInput = React.forwardRef<HTMLInputElement, FilterProps>(({ setFilter
 
   const _onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const v = e.currentTarget.value
+    toggle(true)
     set(v)
   }
 
@@ -134,21 +135,21 @@ export function TitleWithFilter({
         </TitleContainer>
         {!hideCopyButton && (
           <Icon onClick={!copied ? handleCopyClick : undefined} title={`Click to copy all values`}>
-              {!copied ? (
-                <svg xmlns="http://www.w3.org/2000/svg" height="20" viewBox="0 0 20 20" fill="currentColor">
-                  <path d="M8 3a1 1 0 011-1h2a1 1 0 110 2H9a1 1 0 01-1-1z" />
-                  <path d="M6 3a2 2 0 00-2 2v11a2 2 0 002 2h8a2 2 0 002-2V5a2 2 0 00-2-2 3 3 0 01-3 3H9a3 3 0 01-3-3z" />
-                </svg>
-              ) : (
-                <svg xmlns="http://www.w3.org/2000/svg" height="20" viewBox="0 0 20 20" fill="currentColor">
-                  <path d="M9 2a1 1 0 000 2h2a1 1 0 100-2H9z" />
-                  <path
-                    fillRule="evenodd"
-                    d="M4 5a2 2 0 012-2 3 3 0 003 3h2a3 3 0 003-3 2 2 0 012 2v11a2 2 0 01-2 2H6a2 2 0 01-2-2V5zm9.707 5.707a1 1 0 00-1.414-1.414L9 12.586l-1.293-1.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
-                    clipRule="evenodd"
-                  />
-                </svg>
-              )}
+            {!copied ? (
+              <svg xmlns="http://www.w3.org/2000/svg" height="20" viewBox="0 0 20 20" fill="currentColor">
+                <path d="M8 3a1 1 0 011-1h2a1 1 0 110 2H9a1 1 0 01-1-1z" />
+                <path d="M6 3a2 2 0 00-2 2v11a2 2 0 002 2h8a2 2 0 002-2V5a2 2 0 00-2-2 3 3 0 01-3 3H9a3 3 0 01-3-3z" />
+              </svg>
+            ) : (
+              <svg xmlns="http://www.w3.org/2000/svg" height="20" viewBox="0 0 20 20" fill="currentColor">
+                <path d="M9 2a1 1 0 000 2h2a1 1 0 100-2H9z" />
+                <path
+                  fillRule="evenodd"
+                  d="M4 5a2 2 0 012-2 3 3 0 003 3h2a3 3 0 003-3 2 2 0 012 2v11a2 2 0 01-2 2H6a2 2 0 01-2-2V5zm9.707 5.707a1 1 0 00-1.414-1.414L9 12.586l-1.293-1.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
+                  clipRule="evenodd"
+                />
+              </svg>
+            )}
           </Icon>
         )}
         {filterEnabled && (
@@ -165,7 +166,7 @@ export function TitleWithFilter({
         )}
       </StyledTitleWithFilter>
       <FilterWrapper toggled={filterShown}>
-        <FilterInput ref={inputRef} setFilter={setFilter} />
+        <FilterInput ref={inputRef} setFilter={setFilter} toggle={toggle} />
       </FilterWrapper>
     </>
   )