+++ packages/leva/src/components/Monitor/StyledMonitor.ts
@@ -1,8 +1,8 @@
 import { styled } from '../../styles'
 
 export const Canvas = styled('canvas', {
-  height: '$leva__monitorHeight',
+  height: '$monitorHeight',
   width: '100%',
   display: 'block',
-  borderRadius: '$leva__sm',
+  borderRadius: '$sm',
 })