+++ packages/leva/src/types/internal.ts
@@ -1,9 +1,20 @@
 import type { UseStore } from 'zustand'
-import type { SpecialInput, RenderFn, FolderSettings, Plugin } from './public'
+import type { SpecialInput, RenderFn, FolderSettings, Plugin, OnChangeHandler } from './public'
 
 export type State = { data: Data }
 
-export type MappedPaths = Record<string, { path: string; onChange: (value: any) => void; transient: boolean }>
+export type MappedPaths = Record<
+  string,
+  {
+    path: string
+    onChange?: OnChangeHandler
+    onEditStart?: (...args: any) => void
+    onEditEnd?: (...args: any) => void
+    transient: boolean
+  }
+>
+
+type Dispose = () => void
 
 export type StoreType = {
   useStore: UseStore<State>
@@ -16,13 +27,18 @@ export type StoreType = {
   getFolderSettings: (path: string) => FolderSettings
   getData: () => Data
   addData: (newData: Data, override: boolean) => void
-  setValueAtPath: (path: string, value: any) => void
+  setValueAtPath: (path: string, value: any, fromPanel: boolean) => void
   setSettingsAtPath: (path: string, settings: any) => void
   disableInputAtPath: (path: string, flag: boolean) => void
   // TODO possibly better type this
-  set: (values: Record<string, any>) => void
+  set: (values: Record<string, any>, fromPanel: boolean) => void
+  getInput: (path: string) => DataInput | undefined
   get: (path: string) => any
   getDataFromSchema: (schema: any) => [Data, MappedPaths]
+  subscribeToEditStart: (path: string, listener: (value: any) => void) => Dispose
+  subscribeToEditEnd: (path: string, listener: (value: any) => void) => Dispose
+  emitOnEditStart: (path: string) => void
+  emitOnEditEnd: (path: string) => void
 }
 
 export type CommonOptions = {
@@ -35,13 +51,22 @@ export type CommonOptions = {
 export type DataInputOptions = CommonOptions & {
   optional: boolean
   disabled: boolean
-  onChange?: (value: unknown) => void
+}
+
+export type PanelInputOptions = {
+  onChange?: (...args: any) => void
+  onEditStart?: (...args: any) => void
+  onEditEnd?: (...args: any) => void
 }
 
 export type DataInput = {
   __refCount: number
   type: string
   value: unknown
+  /**
+   * Whether the onChange handler invocation is caused internally via the panel or  externally via a set call.
+   */
+  fromPanel: boolean
   settings?: object
 } & DataInputOptions